requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
httpx==0.27.2
python-telegram-bot==21.4
pydantic==2.8.2
pytest==8.3.2
pytest-asyncio==0.23.8
